pyproject.toml
[project]
name = "jason"
version = "0.1.0"
description = "YouTube growth engine for @babygiulybaby — outlier-driven title generation"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "Lucas Barbosa", email = "lucasbarbosa_182@hotmail.com" },
]
dependencies = [
    "typer>=0.12",
    "pydantic>=2.8",
    "pydantic-settings>=2.4",
    "duckdb>=1.0",
    "structlog>=24.4",
    "httpx>=0.28.1",
    "anthropic>=0.39",
    "streamlit>=1.39",
    "google-auth>=2.34",
    "google-auth-oauthlib>=1.2",
    "google-api-python-client>=2.140",
    "fastapi>=0.136.1",
    "uvicorn>=0.46.0",
]

[project.scripts]
jason = "jason.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/jason"]

[dependency-groups]
dev = [
    "pytest>=8.3",
    "respx>=0.23.1",
    "ruff>=0.6",
]
# Optional ML stack (transcription, embeddings, etc.). Install with:
#   uv sync --group ml
ml = [
    "faster-whisper>=1.1",
    "sentence-transformers>=3.0",   # paraphrase-multilingual-mpnet-base-v2 for titles
    "open_clip_torch>=2.26",        # ViT-B-32 for thumbnails
    "pillow>=10.4",                 # required by open_clip preprocessing
    "torch>=2.4",                   # backbone of both
    "bertopic>=0.16",               # two-layer topic modeling (themes + franchises)
    "lightgbm>=4.5",                # multiplier regressor (Fase 3)
    "scikit-learn>=1.5",            # k-means cluster of embeddings + Spearman / metrics
    "pandas>=2.2",                  # feature matrix assembly + temporal split
    "opencv-python>=4.10",          # frame quality + face detection (Fase 4.5)
    "optuna>=4.8.0",
    "pysentimiento>=0.7.3",
    "textstat>=0.7.13",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
# Typer's idiom is `typer.Option(...)` in argument defaults; flagging it as
# B008 is a false positive for every CLI command in this file.
"src/jason/cli.py" = ["B008"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
